refactor(frontend): split PositionDetailDrawer into sub-components (Closes #217)
Extract QuotePanel, KlinePanel, DecisionChainPanel, PositionLocks, and
DetailSection into frontend/web/src/components/position-detail/ to reduce
PositionDetailDrawer.jsx from 513 lines of mixed concerns to a focused
container that manages drawer state, data fetching, and layout composition.
All SSE logic, event cleanup, styling, and props are preserved unchanged.
