+++ frontend/web/src/components/position-detail/KlinePanel.jsx
@@ -1,0 +1,3 @@
+// Re-export KlineChart as KlinePanel for use within position-detail sub-components.
+// KlineChart lives at the shared components level and is already a standalone SVG chart.
+export { default } from '../KlineChart'
