+++ frontend/web/src/components/position-detail/index.jsx
@@ -1,0 +1,1 @@
+export { default } from '../PositionDetailDrawer'
